chore: fmt

+++ test/forge/ModularBundlerTest.sol
@@ -54,7 +54,6 @@ contract ModularBundlerTest is LocalTest {
 
         callbackBundle2.push(_moduleCall(address(module2), abi.encodeCall(module2.isProtected, ())));
 
-
         callbackBundle.push(
             _moduleCall(address(module2), abi.encodeCall(module.callbackBundler, abi.encode(callbackBundle2)))
         );
@@ -65,7 +64,6 @@ contract ModularBundlerTest is LocalTest {
 
         bundle.push(_moduleCall(address(module), abi.encodeCall(module.callbackBundler, abi.encode(callbackBundle))));
 
-
         vm.prank(initiator);
 
         vm.recordLogs();
@@ -74,19 +72,18 @@ contract ModularBundlerTest is LocalTest {
 
         assertEq(entries.length, 8);
 
-        for (uint i = 0; i < entries.length;i++)  {
+        for (uint256 i = 0; i < entries.length; i++) {
             assertEq(entries[i].topics[0], keccak256("CurrentModule(address)"));
         }
 
-        assertEq(entries[0].data,abi.encode(module));
-        assertEq(entries[1].data,abi.encode(module2));
-        assertEq(entries[2].data,abi.encode(module2));
-        assertEq(entries[3].data,abi.encode(module2));
-        assertEq(entries[4].data,abi.encode(module3));
-        assertEq(entries[5].data,abi.encode(module2));
-        assertEq(entries[6].data,abi.encode(module3));
-        assertEq(entries[7].data,abi.encode(module));
-
+        assertEq(entries[0].data, abi.encode(module));
+        assertEq(entries[1].data, abi.encode(module2));
+        assertEq(entries[2].data, abi.encode(module2));
+        assertEq(entries[3].data, abi.encode(module2));
+        assertEq(entries[4].data, abi.encode(module3));
+        assertEq(entries[5].data, abi.encode(module2));
+        assertEq(entries[6].data, abi.encode(module3));
+        assertEq(entries[7].data, abi.encode(module));
     }
 
     function testCurrentModuleSlot() public pure {
